functions/google_finance_price/requirements.txt
functions-framework
requests
beautifulsoup4
pandas
google-cloud-bigquery>=3.12
pytz
